End to end document generation.
From ODT file to a pdf, using the original Typesetr's template info.
Pipeline looks as follows:
1) Parse the initial document using the Typesetr's preprocessor
2) Optimize the structure
3) Write the optimized representation
4) Feed 3) into pandoc to generate a generic target output document
5) Apply Typesetr's templates
6) (if necessary) generate the target format (such as pdf from latex)

MetaInfo entries still need to be correctly parsed and latexified
(some still are represented as FIXME). But the pdf is now generated
correctly.

In order to perform step 6) we copy over the necessary styles into
a temporary directly and run xetex from there.