Extracting meta info from the body of the document
As it turned out some meta information, like title, subtitle
are inferred from the documents directly. This theorotically
meant parsing the document again.
Instead that information is inferred after the document representation
is optimized.

Unfortunately it wasn't so trivial to add that piece of logic
to the existing implementation (i.e. while at the same time
avoiding a lot of mutability) so some chunks of the parser,
optimizer, node extraction functions had to be refactored.
A lot of stuff is now typeclass based, so that when we add support
for more formats you just provide an additional format-specific
type class, if any.

Also added a lot of documentation. Still wip.