Split meta info into schemas and entries
Information loaded from metadata.yml is now treated
as a schema that is used to validate the values
inferred from the document.
Note that a value of the entry, if any, can come from
1) an inferred value
2) the default value specified in the schema
3) the default value for the given type of the entry

Since we now include all meta-entries, the final document
is much closer to what the original typesetr spit out.