On-the-fly addition of styles
During parsing we might want to adapt the initial formatting.
In ODF one cannot simply inline the style but has to define it
in the header of the content.xml. The current way of creating
is through a trial/error so it will need to be adjusted.

Also, while the generated .odt file has the formatting, pandoc
simply ignores it so we have to find a different way for creating
blockquotes.