Post-pandoc fixes to the produces .tex docs
Pandoc does not understand (nor should) Typesetr's `quoting`
environment. What this means is that we can a) try to emulate
it through `quote` and `emph` or b) replace `quote` by `quoting`
in the final document.
a) seems like making `quote` to include italic style inside can
be an extremely tidious task so we go with b) at the risk that
legitimate `quote` get translated into `quoting`. It *looks* like
none of the latter happen (we might have to revisit the fix otherwise)