Workaround Pandoc's missing features
Pandoc does not support some of the formatting issues
that the old typesetr was relying a lot. For example
it has no concept of center or right alignment of text
outside of the table context.

We have two mechanisms as a workaround:
1) redefine environments/commands (in Latex) in the header
of the template. Since we reuse the old templates, the former
are done programatically
2) use custom tags that are then replaced appropriately in
the output writers (typeclass-based)

Citation mechanism in the quotes is done in the optimizer,
rather than the parser. It is driven by the style information.
This is to avoid writing similar logic for each of the parsers.
This is similar to what old typesetr was doing.