Produce pandoc-compatible header info
Pandoc looks into the 'text:outline-level' in order to determine
the level of the header/section. Apparently that is broken in the
odt files (or at least those produced by gdoc) because that information
is enclosed in a cryptic way in the name of the parent style of the
element.
Since we extract the latter, we can also fix the odt document that
we produce.
Shiny section headers ahead!