Local links/references
Pandoc introduces local "pretty-printed" anchors for local links
in the ODT parser. Except that those are broken (there does not appear
to be any kind of state tracking that would ensure that the same
identifiers get the same anchor). Workaround is to only due to translation
for the local bookmarks to references because the latter don't get renamed.