Infer captions and associate them with figures
The optimizer will group together paragrapsh and look at them
in a sliding window of size (currently) 2 fashion. A paragraph
containing a figure, followed by a paragraph with a caption
command will bind into one.

Figures with captions are currently processed by Pandoc, no typesetr
postprocessing is done atm. Depending on the needs we might change that
but the current 'hack' with wrapping figures in text blocks is a bit
cumbersome in this case. Maybe we should introduce paragraphs with text
blocks.