Remove non-determinism in tests
It looked as if tests were sharing the instance of the
singleton object 'Markers' that was wrapping doc fragments
into uniquely named typestr blocks. Except that the counter used
for that would randomly bump dependening on the phase of the moon.
Reseting the counter for each of the tests didn't help, meaning that
there were some concurrency issues there.

Instead we now pass around the token used for generating the unique
names and no non-determinism is present anymore.