Workaround for images in tables
While the produced .odt file is fine, it seems that
.tex doesn't like \tystrblockfigure command when placed
inside the table. So we just inline the command.

The image right now doesn't look pretty because of some
positioning/size issues so we will have to come back to it
and potentially use techniques that are more typesetr-specific.

Relies on Pandoc PR: https://github.com/jgm/pandoc/pull/3218