Improve parser for the formatted text
Previously a text formatted in different ways (e.g. nested
in bold and italic) ended up with multiple layers of unnecessary
<span> blocks, or even wrongly formatted elements. This commit
first of all ads the necessary test to check that the correct
latex is generated in those cases (it wasn't).
Previously we combined together elements of formatted text even
if their styles weren't the same. This is now corrected.